requirements.txt
python-telegram-bot
flask[async]
langchain-core
python-dotenv